+++ docs/modules/is/alpha.ts.md
@@ -1,0 +1,32 @@
+---
+title: is/alpha.ts
+nav_order: 1
+parent: Modules
+---
+
+## alpha overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+- [Is](#is)
+  - [isAlpha](#isalpha)
+
+---
+
+# Is
+
+## isAlpha
+
+check if it's alpha
+
+**Signature**
+
+```ts
+export declare const isAlpha: (a: any) => Right<any> | Left<any>
+```
+
+Added in v0.0.1
