+++ docs/modules/is/alphanumeric.ts.md
@@ -1,0 +1,32 @@
+---
+title: is/alphanumeric.ts
+nav_order: 2
+parent: Modules
+---
+
+## alphanumeric overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+- [Is](#is)
+  - [isAlphanumeric](#isalphanumeric)
+
+---
+
+# Is
+
+## isAlphanumeric
+
+check if it's alphanumeric
+
+**Signature**
+
+```ts
+export declare const isAlphanumeric: (a: any) => Right<any> | Left<any>
+```
+
+Added in v0.0.1
