+++ docs/modules/is/array.ts.md
@@ -1,0 +1,32 @@
+---
+title: is/array.ts
+nav_order: 3
+parent: Modules
+---
+
+## array overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+- [Is](#is)
+  - [isArray](#isarray)
+
+---
+
+# Is
+
+## isArray
+
+check if it's array
+
+**Signature**
+
+```ts
+export declare const isArray: (a: any) => Left<any> | Right<any[]>
+```
+
+Added in v0.0.1
