+++ docs/modules/is/index.ts.md
@@ -1,0 +1,15 @@
+---
+title: is/index.ts
+nav_order: 4
+parent: Modules
+---
+
+## index overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+---
