+++ docs/modules/is/matchRe.ts.md
@@ -1,0 +1,32 @@
+---
+title: is/matchRe.ts
+nav_order: 5
+parent: Modules
+---
+
+## matchRe overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+- [Is](#is)
+  - [isMatchRe](#ismatchre)
+
+---
+
+# Is
+
+## isMatchRe
+
+check if it's match regex
+
+**Signature**
+
+```ts
+export declare const isMatchRe: (re: RegExp) => (a: any) => Right<any> | Left<any>
+```
+
+Added in v0.0.1
