+++ docs/modules/is/notNil.ts.md
@@ -1,0 +1,32 @@
+---
+title: is/notNil.ts
+nav_order: 6
+parent: Modules
+---
+
+## notNil overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+- [Is](#is)
+  - [isNotNil](#isnotnil)
+
+---
+
+# Is
+
+## isNotNil
+
+check if it's not nil
+
+**Signature**
+
+```ts
+export declare const isNotNil: (a: any) => Right<any> | Left<any>
+```
+
+Added in v0.0.1
