+++ docs/modules/is/number.ts.md
@@ -1,0 +1,32 @@
+---
+title: is/number.ts
+nav_order: 7
+parent: Modules
+---
+
+## number overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+- [Is](#is)
+  - [isNumber](#isnumber)
+
+---
+
+# Is
+
+## isNumber
+
+check if it's number
+
+**Signature**
+
+```ts
+export declare const isNumber: (a: any) => Left<any> | Right<number>
+```
+
+Added in v0.0.1
