+++ docs/modules/is/string.ts.md
@@ -1,0 +1,32 @@
+---
+title: is/string.ts
+nav_order: 8
+parent: Modules
+---
+
+## string overview
+
+Added in v0.0.1
+
+---
+
+<h2 class="text-delta">Table of contents</h2>
+
+- [Is](#is)
+  - [isString](#isstring)
+
+---
+
+# Is
+
+## isString
+
+check if it's string
+
+**Signature**
+
+```ts
+export declare const isString: (a: any) => Left<any> | Right<string>
+```
+
+Added in v0.0.1
