+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./String";
 export * from "./Object";
 export * from "./Currency";
+export * from "./is";
