+++ src/internal/index.ts
@@ -1,4 +1,3 @@
-export * from "./is";
 export * from "./function";
 export * from "./stringUtils";
 export * from "./unicodeWords";
