+++ src/is/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @since 0.0.1
+ */
+export * from "./alpha";
+export * from "./alphanumeric";
+export * from "./array";
+export * from "./matchRe";
+export * from "./notNil";
+export * from "./number";
+export * from "./string";
